voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "1.0.0"
description = "LiveKit Voice Agent for Voice AI"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.0.0",
    "livekit-plugins-deepgram>=1.0.0",
    "livekit-plugins-openai>=1.0.0",
    "livekit-plugins-cartesia>=1.0.0",
    "livekit-plugins-silero>=1.0.0",
    "openai>=1.60.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.0.0",
]
